Fixed Cyton start/stop byte encoding

+++ src/libraries/js/src/EventRouter.js
@@ -58,7 +58,7 @@ export class EventRouter{
     }
 
     deinit = () => {
-        this.ui.deleteNode()
+        if (this.ui) this.ui.deleteNode()
     }
 
     // Route Events to Atlas
